cambios1.2

+++ Ejercicios_II_Unidad/Tarea_8_RaizNumero/raizcuadrada.cpp
@@ -6,8 +6,10 @@ using namespace std;
 //Cree una función que calcule la raiz cuadrada de un número 
 //dado utilizando alguna función de la libreria math.h
 
+
+
 double raizCuadrada = 0;
-int raizNumero(double numero){
+double raizNumero(double numero){
 
     //double numero = 4;
     //raizCuadrada = sqrt(numero);
